Updated tests in GenerateAssertionsOnKnownTestCases, replacing use of GetAssertionResponse as GenerateAssertionsService::generateAssertions now returns a RaireResult.

+++ src/main/java/au/org/democracydevelopers/raireservice/persistence/entity/Contest.java
@@ -98,5 +98,5 @@ public class Contest {
    * Get the ID of the contest.
    * @return the contest ID.
    */
-  public int getContestID() { return id; }
+  public long getContestID() { return id; }
 }
